# angular-ui/libs/c4-soft/households-api/model/pageHouseholdDto.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Pageable } from './pageable';
import { HouseholdDto } from './householdDto';
import { Sort } from './sort';


export interface PageHouseholdDto { 
    totalPages?: number;
    totalElements?: number;
    number?: number;
    size?: number;
    content?: Array<HouseholdDto>;
    sort?: Sort;
    first?: boolean;
    last?: boolean;
    numberOfElements?: number;
    pageable?: Pageable;
    empty?: boolean;
}