# angular-ui/libs/c4-soft/households-api/model/pageable.ts
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Sort } from './sort';


export interface Pageable { 
    offset?: number;
    sort?: Sort;
    pageSize?: number;
    pageNumber?: number;
    unpaged?: boolean;
    paged?: boolean;
}