# households-openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"https://localhost:4210","description":"Generated server url"}],"paths":{"/households":{"get":{"tags":["household-controller"],"operationId":"getPage","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}},{"name":"householdLabel","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"taxpayerNameOrId","in":"query","required":false,"schema":{"type":"string","default":""}},{"name":"householdTypeLabel","in":"query","required":false,"schema":{"type":"string","default":""}}],"responses":{"404":{"description":"Not Found"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageHouseholdDto"}}}}}}},"/households/{id}":{"get":{"tags":["household-controller"],"operationId":"getById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"404":{"description":"Not Found"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HouseholdDto"}}}}}}},"/households/types":{"get":{"tags":["household-controller"],"operationId":"getAllTypes","responses":{"404":{"description":"Not Found"},"400":{"description":"Bad Request","content":{"*/*":{"schema":{"type":"string"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HouseholdTypeDto"}}}}}}}}},"components":{"schemas":{"Pageable":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/Sort"},"pageSize":{"type":"integer","format":"int32"},"pageNumber":{"type":"integer","format":"int32"},"unpaged":{"type":"boolean"},"paged":{"type":"boolean"}}},"Sort":{"type":"object","properties":{"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"},"empty":{"type":"boolean"}}},"HouseholdDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"label":{"type":"string"},"type":{"type":"string"},"taxpayer":{"$ref":"#/components/schemas/TaxpayerDto"}}},"PageHouseholdDto":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/HouseholdDto"}},"sort":{"$ref":"#/components/schemas/Sort"},"first":{"type":"boolean"},"last":{"type":"boolean"},"numberOfElements":{"type":"integer","format":"int32"},"pageable":{"$ref":"#/components/schemas/Pageable"},"empty":{"type":"boolean"}}},"TaxpayerDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}}},"HouseholdTypeDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"label":{"type":"string"}}}}}}